Organises dirs, resolves type issues
