+++ app/layout.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import store from "./store/store";
+import store from "../store/store";
 import { Provider } from "react-redux";
 import "./styles/index.scss";
