+++ components/HeadingBar.tsx
